fix(#1002) - Removed entityType, added deploy key and updated test cases
