+++ packages/home-spa/.env.example
@@ -1,1 +1,2 @@
 API_URL=
+DEPLOY_URL=
